requirements.txt
celery==5.5.2
docling==2.31.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
langchain-ollama==0.3.2
openai==1.45.0
psycopg[binary]==3.2.7
python-decouple==3.8